#1053: Rename the class to ProfileDeprecations
