+++ src/main/java/com/rultor/agents/daemons/EndsDaemon.java
@@ -43,6 +43,7 @@ import com.jcabi.xml.XML;
 import com.rultor.Time;
 import com.rultor.agents.AbstractAgent;
 import com.rultor.agents.shells.TalkShells;
+import com.rultor.profiles.ProfileDeprecations;
 import com.rultor.spi.Profile;
 import java.io.IOException;
 import java.util.Collection;
@@ -91,7 +92,7 @@ public final class EndsDaemon extends AbstractAgent {
     @Override
     public Iterable<Directive> process(final XML xml) throws IOException {
         final Shell shell = new TalkShells(xml).get();
-        new DeprecationNotice(this.profile).print(shell);
+        new ProfileDeprecations(this.profile).print(shell);
         final String dir = xml.xpath("/talk/daemon/dir/text()").get(0);
         final int exit = new Script("end.sh").exec(xml);
         final Directives dirs = new Directives();
