+++ src/main/java/com/rultor/agents/daemons/StartsDaemon.java
@@ -40,6 +40,7 @@ import com.jcabi.xml.XML;
 import com.rultor.Time;
 import com.rultor.agents.AbstractAgent;
 import com.rultor.agents.shells.TalkShells;
+import com.rultor.profiles.ProfileDeprecations;
 import com.rultor.spi.Profile;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
@@ -115,7 +116,7 @@ public final class StartsDaemon extends AbstractAgent {
     public String run(final XML xml) throws IOException {
         final XML daemon = xml.nodes("/talk/daemon").get(0);
         final Shell shell = new TalkShells(xml).get();
-        new DeprecationNotice(this.profile).print(shell);
+        new ProfileDeprecations(this.profile).print(shell);
         final ByteArrayOutputStream baos = new ByteArrayOutputStream();
         new Shell.Safe(shell).exec(
             "mktemp -d -t rultor-XXXX",
